requirements_qwen3_vl.txt
accelerate==1.11.0
deepspeed==0.13.1
torch==2.3.1
torchvision==0.18.1
matplotlib==3.9.2
numpy==1.23.5
pandas==2.2.3
peft==0.11.1
Pillow
PyYAML
qwen-vl-utils=0.0.14
Requests==2.32.3
scipy==1.14.1
tqdm
transformers=4.57.0
